refactor: updated imgs in projects, button hover style
